Removed warnings from project (duplicated build-copy-phase etc)

+++ WakaLet/AppDelegate.m
@@ -133,7 +133,7 @@
     } else if (button == NSAlertAlternateReturn) {
         return nil;
     } else {
-        NSAssert1(NO, @"Invalid input dialog button %d", button);
+        NSAssert1(NO, @"Invalid input dialog button %lu", (long)button);
         return nil;
     }
 }
